move xml tests to module
