+++ tests/xml/data/hamlet.xml
