+++ tests/xml/test_xml_reader.py
@@ -1,4 +1,40 @@
-from .xml_reader import HamletXMLReader
+import os
+
+from ianalyzer_readers.readers.xml import XMLReader
+from ianalyzer_readers.readers.core import Field
+from ianalyzer_readers.extract import XML
+
+class HamletXMLReader(XMLReader):
+    """
+    Example XML reader for testing
+    """
+
+    data_directory = os.path.join(os.path.dirname(__file__), 'data')
+
+    tag_toplevel = 'document'
+    tag_entry = 'lines'
+
+    def sources(self, **kwargs):
+        for filename in os.listdir(self.data_directory):
+            full_path = os.path.join(self.data_directory, filename)
+            yield full_path, {
+                'filename': filename
+            }
+
+    title = Field(
+        'title',
+        XML('title', toplevel=True, recursive=True)
+    )
+    character = Field(
+        'character',
+        XML(None, attribute='character')
+    )
+    lines = Field(
+        'lines',
+        XML('l', multiple=True, transform='\n'.join),
+    )
+
+    fields = [title, character, lines]
 
 target_documents = [
     {
@@ -44,7 +80,7 @@ target_documents = [
 ]
 
 
-def test_xml():
+def test_xml_reader():
     reader = HamletXMLReader()
     docs = reader.documents()
